Allow RPM-based distro to pass source for package
The Java class currently assumes that a Java package is present
in an available repository.  This commit allows a user to pass
a string indicating a source package location, which will cause
the class to switch to the RPM provider from yum.

No provision is made for Debian, but backwards compatibility is
maintained and it should be straightforward to extend for debs.